backend/pyproject.toml
[project]
name = "stockanalysisassistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.124.4",
    "langchain[openai]>=1.2.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
    "yfinance>=0.2.66",
    "supabase>=2.10.0",
]
